requirements.txt
langchain
langchain-community
langchain-google-genai
google-generativeai
openai
faiss-cpu
chromadb
pypdf
python-docx
streamlit
python-dotenv
tiktoken